requirements.txt
Flask==3.0.0
werkzeug==3.0.1
pdfplumber==0.10.3
google-genai
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
python-dotenv==1.0.0
openpyxl==3.1.2
gunicorn==21.2.0
pandas==2.2.2
numpy<2.0.0
setuptools